+++ event-sourcing/event-publisher/published-event-processor/src/test/java/uk/gov/justice/services/eventsourcing/publishedevent/rebuild/integration/RebuildPerformanceIT.java
@@ -22,6 +22,7 @@ import uk.gov.justice.services.eventsourcing.repository.jdbc.EventInsertionStrat
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.EventConverter;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.EventJdbcRepository;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.eventstream.EventStreamJdbcRepository;
+import uk.gov.justice.services.eventsourcing.util.io.Closer;
 import uk.gov.justice.services.jdbc.persistence.JdbcResultSetStreamer;
 import uk.gov.justice.services.jdbc.persistence.PreparedStatementWrapperFactory;
 import uk.gov.justice.services.messaging.DefaultJsonObjectEnvelopeConverter;
@@ -103,7 +104,8 @@ public class RebuildPerformanceIT {
             LoggerProducer.class,
             DefaultJsonEnvelopeProvider.class,
             DefaultEnvelopeProvider.class,
-            DummySystemCommandStore.class
+            DummySystemCommandStore.class,
+            Closer.class
     })
     @Default
     public WebApp war() {