+++ event-sourcing/event-publisher/published-event-processor/src/test/java/uk/gov/justice/services/eventsourcing/publishedevent/rebuild/integration/RebuildPublishedEventIT.java
@@ -28,6 +28,7 @@ import uk.gov.justice.services.eventsourcing.repository.jdbc.event.EventConverte
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.EventJdbcRepository;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.event.PublishedEvent;
 import uk.gov.justice.services.eventsourcing.repository.jdbc.eventstream.EventStreamJdbcRepository;
+import uk.gov.justice.services.eventsourcing.util.io.Closer;
 import uk.gov.justice.services.jdbc.persistence.JdbcResultSetStreamer;
 import uk.gov.justice.services.jdbc.persistence.PreparedStatementWrapperFactory;
 import uk.gov.justice.services.messaging.DefaultJsonObjectEnvelopeConverter;
@@ -118,7 +119,8 @@ public class RebuildPublishedEventIT {
             ObjectMapperProducer.class,
             LoggerProducer.class,
             DefaultJsonEnvelopeProvider.class,
-            DefaultEnvelopeProvider.class
+            DefaultEnvelopeProvider.class,
+            Closer.class
     })
     @Default
     public WebApp war() {